Fix typo for dequeue and types for arg validation

+++ inc/message.h
@@ -12,8 +12,8 @@ public:
 protected:
     const size_t MAX_STR_SIZE{50};
     const size_t MIN_STR_SIZE{1};
-    const size_t MAX_INT_VAL{1000};
-    const size_t MIN_INT_VAL{1};
+    const int MAX_INT_VAL{1000};
+    const int MIN_INT_VAL{1};
 
     std::string _message_id{};
 };
